Create program to get Process Id and Parent Process Id in Linux
This program will get the Process Id and Parent Process Id of the current Process in C programming Linux.

Here we are using two functions getpid() to get Process Id and getppid() to get Parent Process Id of the current process. These functions are declared in <unistd.h> header file.